refac: Introduce pm promotions service
